v1.0

+++ src/containers/EditNote.js
@@ -1,6 +1,8 @@
 import { connect } from "react-redux";
 import UIComponent from "../components/NoteEdit";
 
+import store from "../store";
+
 import * as notesActions from "../actions/notesActions";
 import * as labelsActions from "../actions/labelsActions";
 import * as editActions from "../actions/currentEditNoteActions";
@@ -13,7 +15,11 @@ const checkLabelExist = (notes, label) => {
 
 const mapStateToProps = state => {
   return state.noteList[state.currentEditNote]
-    ? { ...state.noteList[state.currentEditNote], open: true }
+    ? {
+        ...state.noteList[state.currentEditNote],
+        labelList: state.labels,
+        open: true
+      }
     : { open: false };
 };
 
@@ -24,11 +30,10 @@ const mapDispatchToProps = dispatch => {
       dispatch(notesActions.addLabelAction(id, label));
       dispatch(labelsActions.addLabelAction(label));
     },
-    removeLabel: (id, label) => {
+    removeLabel: (id, label, notes) => {
       dispatch(notesActions.removeLabelAction(id, label));
-      console.log(checkLabelExist(label));
-      // if (!checkLabelExist(label))
-      dispatch(labelsActions.removeLabelAction(label));
+      if (!checkLabelExist(store.getState().noteList, label))
+        dispatch(labelsActions.removeLabelAction(label));
     },
     editHandle: (id, editFields) => {
       dispatch(notesActions.editNoteAction(id, editFields));
